+++ Harbor/plugins/Plugin.js
@@ -1,4 +1,4 @@
-import Core from '../common/Core.js';
+import { Core } from '../common/Core.js';
 
 /**
  * Creates a new Harbor Plugin that will be registered to the TaskManager.
@@ -11,7 +11,7 @@ import Core from '../common/Core.js';
  * @param {Object} options Defines the Harbor specific options for the current
  * service.
  */
-class Plugin extends Core {
+export class Plugin extends Core {
   constructor(services, options, workers) {
     super(services, options, 'plugins', workers);
 
@@ -35,5 +35,3 @@ class Plugin extends Core {
     super.defineEntry(true);
   }
 }
-
-export default Plugin;